Move `HydrogenRequest` out of weird Hydration folder
