+++ packages/hydrogen/src/foundation/session/session.ts
@@ -1,7 +1,7 @@
 import {Logger} from '../../utilities/log';
 import {wrapPromise} from '../../utilities/suspense';
 import type {ServerComponentResponse} from '../../framework/Hydration/ServerComponentResponse.server';
-import type {HydrogenRequest} from '../../framework/Hydration/HydrogenRequest.server';
+import type {HydrogenRequest} from '../../framework/HydrogenRequest.server';
 
 export type SessionSyncApi = {
   get: () => Record<string, string>;
