+++ packages/hydrogen/src/framework/Hydration/tests/HydrogenRequest.test.ts
@@ -1,7 +1,7 @@
 import {IncomingMessage} from 'http';
 import {RSC_PATHNAME} from '../../../constants';
 import {PreloadOptions} from '../../../types';
-import {PreloadQueryEntry, HydrogenRequest} from '../HydrogenRequest.server';
+import {PreloadQueryEntry, HydrogenRequest} from '../../HydrogenRequest.server';
 
 describe('HydrogenRequest', () => {
   it('converts node request to Fetch API request', () => {
