+++ packages/hydrogen/src/utilities/log/log-query-timeline.ts
@@ -1,4 +1,4 @@
-import {HydrogenRequest} from '../../framework/Hydration/HydrogenRequest.server';
+import {HydrogenRequest} from '../../framework/HydrogenRequest.server';
 import {QueryKey} from '../../types';
 import {hashKey} from '../hash';
 import {findQueryName, parseUrl} from './utils';
